requirements.txt
streamlit
huggingface_hub
fpdf2
graphviz
pillow